+++ assets/javascripts/kitten/components/navigation/header-nav/components/dropdown.js
@@ -17,6 +17,7 @@ import {
   keyboardNavigation,
 } from '../../../../helpers/dom/a11y'
 import { DROPDOWN_ANIMATED_DELAY } from '../../../../constants/dropdown-config'
+import { ArrowIcon } from '../../../../components/icons/arrow-icon'
 
 export const Dropdown = React.forwardRef(
   (
@@ -31,6 +32,7 @@ export const Dropdown = React.forwardRef(
       closeOnOuterClick,
       dropdownContent,
       dropdownContentWidth,
+      hasArrow,
       isExpanded,
       keepInitialButtonAction,
       onPositionUpdate,
@@ -316,9 +318,23 @@ export const Dropdown = React.forwardRef(
               onClick={handleButtonClick}
             >
               {button}
+              {hasArrow && (
+                <ArrowIcon
+                  direction={isExpandedState ? 'top' : 'bottom'}
+                  className="k-u-margin-left-single k-u-hidden@xs-down"
+                />
+              )}
             </DropdownButton>
           ) : (
-            button
+            <>
+              {button}
+              {hasArrow && (
+                <ArrowIcon
+                  direction={isExpandedState ? 'top' : 'bottom'}
+                  className="k-u-margin-left-single k-u-hidden@xs-down"
+                />
+              )}
+            </>
           ))}
 
         {!button && (