+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: `HeaderNav`: Add `size` prop, accepts `regular` and `small`.
+- Feature: `HeaderNav.UserMenu`: Add `hasArrow` boolean prop.
 
 ## [3.9.0] - 2021-05-06